separate event listener from plugin class
