+++ src/main/java/com/github/therealguru/totemfletching/TotemFletchingPlugin.java
@@ -1,19 +1,13 @@
 package com.github.therealguru.totemfletching;
 
-import com.github.therealguru.totemfletching.service.EntTrailService;
+import com.github.therealguru.totemfletching.listener.PluginEventListener;
 import com.github.therealguru.totemfletching.service.OverlayService;
-import com.github.therealguru.totemfletching.service.ResearchPointService;
-import com.github.therealguru.totemfletching.service.TotemService;
 import com.google.inject.Provides;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.Client;
-import net.runelite.api.GameState;
 import net.runelite.api.events.*;
-import net.runelite.api.gameval.VarPlayerID;
-import net.runelite.api.gameval.VarbitID;
 import net.runelite.client.config.ConfigManager;
-import net.runelite.client.eventbus.Subscribe;
+import net.runelite.client.eventbus.EventBus;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 
@@ -21,62 +15,24 @@ import net.runelite.client.plugins.PluginDescriptor;
 @PluginDescriptor(name = "Totem Fletching")
 public class TotemFletchingPlugin extends Plugin {
 
-    @Inject private Client client;
-
+    @Inject private EventBus eventBus;
+    @Inject private PluginEventListener pluginEventListener;
     @Inject private OverlayService overlayService;
-    @Inject private TotemService totemService;
-    @Inject private EntTrailService entTrailService;
-    @Inject private ResearchPointService researchPointService;
 
     @Override
     protected void startUp() throws Exception {
+        eventBus.register(pluginEventListener);
         overlayService.registerOverlays();
     }
 
     @Override
     protected void shutDown() throws Exception {
+        eventBus.unregister(pluginEventListener);
         overlayService.unregisterOverlays();
     }
 
     @Provides
     TotemFletchingConfig provideConfig(ConfigManager configManager) {
         return configManager.getConfig(TotemFletchingConfig.class);
-    }
-
-    @Subscribe
-    public void onVarbitChanged(final VarbitChanged varbitChanged) {
-        if (varbitChanged.getVarbitId() == VarPlayerID.ENT_TOTEMS_RESEARCH_POINTS) {
-            researchPointService.onVarbitChanged();
-        }
-        if (varbitChanged.getVarbitId() < VarbitID.ENT_TOTEMS_SITE_1_BASE
-                || varbitChanged.getVarbitId() > VarbitID.ENT_TOTEMS_SITE_8_ALL_MULTIANIMALS)
-            return;
-
-        totemService.onVarbitChanged(varbitChanged);
-    }
-
-    @Subscribe
-    public void onGameObjectSpawned(final GameObjectSpawned gameObjectSpawned) {
-        totemService.addGameObject(gameObjectSpawned.getGameObject());
-        entTrailService.addEntTrail(gameObjectSpawned);
-    }
-
-    @Subscribe
-    public void onGameObjectDespawned(final GameObjectDespawned gameObjectDespawned) {
-        totemService.removeGameObject(gameObjectDespawned.getGameObject());
-        entTrailService.removeEntTrail(gameObjectDespawned);
-    }
-
-    @Subscribe
-    public void onGameStateChanged(GameStateChanged event) {
-        if (event.getGameState().equals(GameState.LOADING)) {
-            totemService.clearGameObjects();
-            entTrailService.clearEntTrails();
-        }
-    }
-
-    @Subscribe
-    public void onGameTick(final GameTick gameTick) {
-        totemService.updateClosestTotem(client.getLocalPlayer());
     }
 }
